refactor: tts.py 拆出配置层与音频检查层 (C9)
- 新增 app/voice/tts_settings.py：GPTSoVITSTTSSettings/ToneReference/
  TTSConfigError、provider 与播放后端常量、语气参考解析（纯配置层，
  不依赖 Qt 与网络）
- 新增 app/voice/audio_checks.py：_verify_generated_audio/_wav_duration_ms/
  _is_valid_wav_file（纯函数检查层）
- tts.py 全量 re-export 拆出符号，旧 import 路径与测试 patch 路径零变化；
  瘦身 2756→2245 行
- 拆分边界刻意保守：被测试 monkeypatch 的符号（time/_probe_*/_find_*/
  进程管理）全部留在 tts.py，避免 patch 路径失效

测试：全量 813 passed / exit 0（零测试改动，验证 re-export 完整性）
